+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Change the datasource url to be the multi-tenant-proxy in front of the loki-gateway.
+
 ## [0.7.3] - 2024-08-13
 
 ### Fixed